moved copy clipboad to component
